+++ src/Extensions/src/extensions/recastjs/recastjs_crowd.cpp
@@ -22,7 +22,7 @@ RecastJSCrowd::RecastJSCrowd(RecastJSPlugin* plugin, size_t maxAgents, float max
   _scene          = scene;
 
   _onBeforeAnimationsObserver
-    = scene->onBeforeAnimationsObservable.add([this](Scene* scene, EventState& /*es*/) -> void {
+    = scene->onBeforeAnimationsObservable.add([this](Scene* scene, EventState & /*es*/) -> void {
         update(scene->getEngine()->getDeltaTime() * 0.001f);
       });
 }
@@ -166,7 +166,7 @@ void RecastJSCrowd::update(float deltaTime)
   else {
     auto iterationCount = deltaTime / timeStep;
     if (maxStepCount && iterationCount > maxStepCount) {
-      iterationCount = maxStepCount;
+      iterationCount = static_cast<float>(maxStepCount);
     }
     if (iterationCount < 1) {
       iterationCount = 1;
